refactor: move index.ts to src, and adjust import paths of it
