--- index.ts
@@ -1,8 +1,0 @@
-export * from './src/common/';
-export * from './src/engine/';
-export * from './src/elements/';
-export * from './src/server/';
-export * from './src/datastore/';
-
-export * from './src/interfaces/';
-
