+++ src/elements/Element.ts
@@ -2,7 +2,7 @@
 import { Execution } from '../engine/Execution';
 import { Token } from '../engine/Token';
 import { IBehaviour, Behaviour} from './behaviours/.';
-import { NODE_ACTION, FLOW_ACTION, EXECUTION_EVENT, TOKEN_STATUS, ITEM_STATUS } from '../../';
+import { NODE_ACTION, FLOW_ACTION, EXECUTION_EVENT, TOKEN_STATUS, ITEM_STATUS } from '..';
 
 import { Item } from '../engine/Item';
 import { Node } from '.';
