+++ src/elements/behaviours/Behaviour.ts
@@ -1,7 +1,7 @@
 import { TimerBehaviour, CamundaFormData, IOBehaviour, MessageEventBehaviour, SignalEventBehaviour, TerminateBehaviour, LoopBehaviour } from ".";
 import  type { Node } from "..";
 import type { Item } from "../../engine/Item";
-import { IItem } from "../../..";
+import { IItem } from "../..";
 import { ScriptBehaviour } from "./Script";
