+++ src/interfaces/engine.ts
@@ -1,6 +1,6 @@
 import { ITEM_STATUS, EXECUTION_STATUS, NODE_ACTION, FLOW_ACTION, TOKEN_STATUS } from './Enums';
 import { IItemData , IInstanceData } from './';
-import { ILogger, IAppDelegate, IBPMNServer, IDefinition, IElement, Execution, Token, Item, Element, INode, Node, IServerComponent } from '../../';
+import { ILogger, IAppDelegate, IBPMNServer, IDefinition, IElement, Execution, Token, Item, Element, INode, Node, IServerComponent } from '..';
 import { EventEmitter } from 'events';
 interface IDataStore {
